==> 0_download_data.ipynb <==
import ee
import geopandas as gpd
from funcs import export_image, getWeeklySentinelComposite

ee.Authenticate()
ee.Initialize()

## Data Acquisition (google earth engine)

project_shp = gpd.read_file('area/study_area.geojson') #read shapefile

bbox = project_shp.total_bounds.tolist()
aoi = ee.Geometry.Rectangle(bbox)

startDate = '2023-10-01'
endDate = '2024-03-31'
data_source = "COPERNICUS/S2_SR_HARMONIZED"
s2_list_size, sen2w_list = getWeeklySentinelComposite(startDate, endDate, data_source, aoi)

# export images to drive
for i in range(s2_list_size):
  export_image(ee.Image(sen2w_list.get(i)), aoi)
  print(f'{i+1} / {s2_list_size} done')
