==> 1_stack_images.ipynb <==
## Read and stack images

import os
from funcs import stack_images

data_dr = os.path.join('data', 's2') # where data is saved
output_dr = os.path.join('data', 'stacked_bands.tif') # where to save the data
stack_images(data_dr, output_dr)
