==> 5_visualize.ipynb <==
import os
import rasterio
import rasterio.mask
import geopandas as gpd
from funcs import Map

map = Map()
class_colors = ['purple', 'green', 'goldenrod', 'brown']
class_names = ['beans', 'potato', 'wheat', 'others']

src = rasterio.open(os.path.join('results', 'classified.tif'))
title = 'Tatti Crop Classification'
map.draw_map(src, class_names, class_colors, title)

# display only test area
aoi = gpd.read_file(os.path.join('area', 'aoi.geojson'))
src = rasterio.open(os.path.join('results', 'classified.tif'))
croped_array, transform = rasterio.mask.mask(src, aoi.geometry, crop=True)
meta = src.meta
meta.update(width= croped_array.shape[2], height= croped_array.shape[1], transform= transform)
with rasterio.open(os.path.join('results', 'classified_aoi.tif'), 'w', **meta) as dst:
	dst.write(croped_array)

src = rasterio.open(os.path.join('results', 'classified_aoi.tif'))
title = 'Tatti Crop Classification (AOI)'
map.draw_map(src, class_names, class_colors, title)
